# api/tsoa/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ICategory": {
				"properties": {
					"categoryName": {
						"type": "string"
					},
					"categoryDescription": {
						"type": "string"
					},
					"categoryImageUrl": {
						"type": "string"
					}
				},
				"required": [
					"categoryName",
					"categoryDescription",
					"categoryImageUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProduct": {
				"properties": {
					"categoryName": {
						"type": "string"
					},
					"productName": {
						"type": "string"
					},
					"previewImageUrl": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"postedDate": {
						"type": "string",
						"format": "date-time"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"isAvailable": {
						"type": "boolean"
					},
					"discountPercentage": {
						"type": "number",
						"format": "double"
					},
					"discountEndDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"categoryName",
					"productName",
					"previewImageUrl",
					"description",
					"postedDate",
					"price",
					"isAvailable",
					"discountPercentage",
					"discountEndDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRoutes": {
				"properties": {
					"path": {
						"type": "string"
					},
					"stack": {
						"items": {
							"properties": {
								"method": {
									"type": "string"
								},
								"regexp": {
									"properties": {
										"fast_slash": {
											"type": "boolean"
										},
										"fast_star": {
											"type": "boolean"
										}
									},
									"required": [
										"fast_slash",
										"fast_star"
									],
									"type": "object"
								},
								"keys": {
									"items": {
										"type": "string"
									},
									"type": "array"
								},
								"name": {
									"type": "string"
								}
							},
							"required": [
								"method",
								"regexp",
								"keys",
								"name"
							],
							"type": "object"
						},
						"type": "array"
					},
					"methods": {
						"properties": {},
						"additionalProperties": {
							"type": "boolean"
						},
						"type": "object"
					}
				},
				"required": [
					"path",
					"stack",
					"methods"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": ""
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/categories": {
			"get": {
				"operationId": "GetAllCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICategory"
							}
						}
					}
				}
			}
		},
		"/api/categories/{categoryName}": {
			"get": {
				"operationId": "GetCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryName",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/categories/{id}": {
			"put": {
				"operationId": "UpdateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICategory"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/products": {
			"get": {
				"operationId": "GetAllProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProduct"
							}
						}
					}
				}
			}
		},
		"/api/products/{productName}": {
			"get": {
				"operationId": "GetProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "productName",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/products/category/{categoryName}": {
			"get": {
				"operationId": "GetAllProductsFromACategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryName",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/products/{id}": {
			"put": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProduct"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/users": {
			"get": {
				"operationId": "GetAllUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"roles": {
										"items": {
											"type": "string"
										},
										"type": "array"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"roles",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/users/{name}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/users/{id}": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"roles": {
										"items": {
											"type": "string"
										},
										"type": "array"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"roles",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/server-status": {
			"get": {
				"operationId": "GetServerStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"serverTime": {
											"type": "string"
										},
										"status": {
											"type": "string"
										}
									},
									"required": [
										"serverTime",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Server Status"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/server-status/routes": {
			"get": {
				"operationId": "GetServerRoutes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IRoutes"
								}
							}
						}
					}
				},
				"tags": [
					"Server Status"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}